SumOfArrayof2D
In this program, I solve the problem of the sum of an array.
and I take two arrays one is A[][] and another is B[][] I take two DS. First I take input from the user and then take the sum of both and then display them I used them the for iteration loop for the input, sum, and display.